Performced one navigation run
